+++ examples/versions.tf
@@ -1,10 +1,16 @@
 
 terraform {
-  required_version = "~> 0.12.0"
+  required_version = "~> 0.13.0"
   required_providers {
     local    = "~> 1.2"
     template = "~> 2.1"
-    ct = "~> 0.5.0"
+
+    ct = {
+      source  = "poseidon/ct"
+      version = "~> 7.0.0"
+      #source = "terraform.localhost/poseidon/ct"
+      #version = "0.7.0"
+    }
   }
 }